Chore: split dashboard thumbnail service
