+++ pkg/services/dashboard_thumbs/dashboardthumbsimpl/dashboard_thumbs.go
@@ -5,24 +5,23 @@ import (
 
 	"github.com/grafana/grafana/pkg/models"
 	dashboardthumbs "github.com/grafana/grafana/pkg/services/dashboard_thumbs"
-	"github.com/grafana/grafana/pkg/services/sqlstore"
+	"github.com/grafana/grafana/pkg/services/sqlstore/db"
 )
 
 type Service struct {
-	// TODO remove sqlstore
-	sqlStore *sqlstore.SQLStore
+	store store
 }
 
 func ProvideService(
-	ss *sqlstore.SQLStore,
+	db db.DB,
 ) dashboardthumbs.Service {
 	return &Service{
-		sqlStore: ss,
+		store: &xormStore{db: db},
 	}
 }
 
 func (s *Service) GetThumbnail(ctx context.Context, query *models.GetDashboardThumbnailCommand) (*models.DashboardThumbnail, error) {
-	dt, err := s.sqlStore.GetThumbnail(ctx, query)
+	dt, err := s.store.Get(ctx, query)
 	if err != nil {
 		return dt, err
 	}
@@ -30,7 +29,7 @@ func (s *Service) GetThumbnail(ctx context.Context, query *models.GetDashboardTh
 }
 
 func (s *Service) SaveThumbnail(ctx context.Context, cmd *models.SaveDashboardThumbnailCommand) (*models.DashboardThumbnail, error) {
-	dt, err := s.sqlStore.SaveThumbnail(ctx, cmd)
+	dt, err := s.store.Save(ctx, cmd)
 	if err != nil {
 		return dt, err
 	}
@@ -38,7 +37,7 @@ func (s *Service) SaveThumbnail(ctx context.Context, cmd *models.SaveDashboardTh
 }
 
 func (s *Service) UpdateThumbnailState(ctx context.Context, cmd *models.UpdateThumbnailStateCommand) error {
-	err := s.sqlStore.UpdateThumbnailState(ctx, cmd)
+	err := s.store.UpdateState(ctx, cmd)
 	if err != nil {
 		return err
 	}
@@ -46,7 +45,7 @@ func (s *Service) UpdateThumbnailState(ctx context.Context, cmd *models.UpdateTh
 }
 
 func (s *Service) FindThumbnailCount(ctx context.Context, cmd *models.FindDashboardThumbnailCountCommand) (int64, error) {
-	i, err := s.sqlStore.FindThumbnailCount(ctx, cmd)
+	i, err := s.store.Count(ctx, cmd)
 	if err != nil {
 		return i, err
 	}
@@ -54,7 +53,7 @@ func (s *Service) FindThumbnailCount(ctx context.Context, cmd *models.FindDashbo
 }
 
 func (s *Service) FindDashboardsWithStaleThumbnails(ctx context.Context, cmd *models.FindDashboardsWithStaleThumbnailsCommand) ([]*models.DashboardWithStaleThumbnail, error) {
-	d, err := s.sqlStore.FindDashboardsWithStaleThumbnails(ctx, cmd)
+	d, err := s.store.FindDashboardsWithStaleThumbnails(ctx, cmd)
 	if err != nil {
 		return d, err
 	}
